+++ pyoxidizer/src/project_building.rs
@@ -156,6 +156,12 @@ pub fn build_executable_with_rust_project<'a>(
     if exe.requires_jemalloc() {
         features.push("jemalloc");
     }
+    if exe.requires_mimalloc() {
+        features.push("mimalloc");
+    }
+    if exe.requires_snmalloc() {
+        features.push("snmalloc");
+    }
 
     let features = features.join(" ");