+++ pyoxidizer/src/py_packaging/binary.rs
@@ -338,6 +338,12 @@ pub trait PythonBinaryBuilder {
     /// Whether the binary requires the jemalloc library.
     fn requires_jemalloc(&self) -> bool;
 
+    /// Whether the binary requires the Mimalloc library.
+    fn requires_mimalloc(&self) -> bool;
+
+    /// Whether the binary requires the Snmalloc library.
+    fn requires_snmalloc(&self) -> bool;
+
     /// Obtain an `EmbeddedPythonContext` instance from this one.
     fn to_embedded_python_context(
         &self,