+++ python-packaging/src/interpreter.rs
@@ -101,6 +101,10 @@ pub enum MemoryAllocatorBackend {
     System,
     /// Use jemalloc.
     Jemalloc,
+    /// Use Mimalloc.
+    Mimalloc,
+    /// Use Snmalloc.
+    Snmalloc,
     /// Use Rust's global allocator.
     Rust,
 }
@@ -110,6 +114,8 @@ impl ToString for MemoryAllocatorBackend {
         match self {
             Self::System => "system",
             Self::Jemalloc => "jemalloc",
+            Self::Mimalloc => "mimalloc",
+            Self::Snmalloc => "snmalloc",
             Self::Rust => "rust",
         }
         .to_string()
@@ -123,6 +129,8 @@ impl TryFrom<&str> for MemoryAllocatorBackend {
         match value {
             "system" => Ok(Self::System),
             "jemalloc" => Ok(Self::Jemalloc),
+            "mimalloc" => Ok(Self::Mimalloc),
+            "snmalloc" => Ok(Self::Snmalloc),
             "rust" => Ok(Self::Rust),
             _ => Err(format!("{} is not a valid memory allocator backend", value)),
         }
@@ -153,6 +161,20 @@ impl PythonRawAllocator {
     pub fn jemalloc() -> Self {
         Self {
             backend: MemoryAllocatorBackend::Jemalloc,
+            ..PythonRawAllocator::default()
+        }
+    }
+
+    pub fn mimalloc() -> Self {
+        Self {
+            backend: MemoryAllocatorBackend::Mimalloc,
+            ..PythonRawAllocator::default()
+        }
+    }
+
+    pub fn snmalloc() -> Self {
+        Self {
+            backend: MemoryAllocatorBackend::Snmalloc,
             ..PythonRawAllocator::default()
         }
     }